For non-defining opaque type usage errors, don't try to also prove all trait bounds
